[Nova] Enable token caching for openstack-agent-liveness
With [0] the `python-agentliveness` command got the ability to cache the
token and re-use it. We would like to use that functionality, because
requesting a new token every liveness check a) puts more load on
keystone for getting the token and b) puts more load on keystone for the
service not having cached the token and thus asking keystone for its
validity.

[0] https://github.com/sapcc/python-agentliveness/pull/7