[Neutron] Explicitly set user for NSX-T password generation
nsxv3-agents use the `vcenter-operator` to get deployed. The
`vcenter-operator` is responsible for generating the individual agent's
password with the masterpassword algorithm. While we use the user
"admin" to log into the NSX-T agents, we used the current
`vcenter-operator`'s user to generate the password.

With changing the `vcenter-operator`'s user to using a user from the
`vsphere.local` domain, we now can't use the from there anymore.
Instead, we introduce a new variable to read the user from and set the
old user there.