[rabbitmq] add sidecar container as configurator
add a sidecar container to configure rabbitmq access credentials without
the need to reboot rabbitmq and keep sure that the user configuration
are rolled out correctly.
this feature adds a readiness check to check the compliance to the new
configuration.
the feature presents a new value `configurator.deletionForceType` which
defines the deletion behavior:
1. force : old users will be removed immediately
2. gracefully: wait for  all user's connection to be closed before
        deleting it, and fail after `configurator.deletionTimeout`
3. force_gracefully : wait for  all user's connection to be closed
   before deleting, force deletion after `configurator.deletionTimeout`

Change-Id: If851efe346078b4e2065f5b6e4e2d604572083f0