Remove complicated call to templates with tuples
We called templates by creating a tuple of all necessary arguments for
the template. This lead to long calls and a cluttered chart.

Let us instead create 2 dicts. One scoped to the domain, one that
contains additional values for the instance. Both dicts also get the top
level context, so we can access .Values, .Release and so on.

We then pass these dicts to the relevant templates (depending if the
template is instance-scoped or domain scoped). This removes also the
long headers translating the tuple back into vars at the template's
head.

This is a transparent change, the resulting chart should not change.