Let router-id be determined from init-network
The router id is a 4 byte value that is written in IPv4 notation. Bird
used the line `router id from "vlan*";` meaning it takes the IP address
from an interface starting with vlan* as router id. With v6 that does
not work anymore because the router id is still the same value.

There are 2 options:
1. hardcode the router id in the config
2. Dynamically determine router id at runtime

(1) has the disadvantage that we need a dedicated configmap per route
server instance, meaning we would also back out from the plan that we
run both router-server instances from a single statefulset as 2
configmaps need 2 statefulsets. It would also mean, that the
configbuilder, which right now has no knowledge of the IP addresses bird
is served from must get this information, likely by getting a
secrets-repo reference or by it bying replicated. This is all not needed
for (2), the only downside would be that some parts of the config cannot
be validated upon building/merging/deploying it. We would create an
additional config that is referenced from the main config that only
contains a single line, the router id. As the router id will be the
corrresponding router-servers v4 address we always feed the v4 address
into the pod, even if it is a v6 pod. We let the init-network script
calculate the link IP and the router id. It will generate a single line
into a file that the bird config will include.