[CINDER] update fcd name and add quality type
This patch updates the name of the cinder backend definition for
cinder-volume.conf for the fcd driver and adds the proper
storage profile setting.

This also adds the new extra_capability reporting for each backend
called quality_type.  For vmware and fcd the quality type is set
to premium.  For standard_hdd the quality type is set to standard.

This allows us to add volume type extra specs to distinguish
on backend based on quality_type.