ironic: enable watcher only in API pods
This puts the watcher config in it's own config file below ironic.conf.d
That allows us to only include the watcher config in API deployments.
The conductors don't need the watcher config but had it via the shared
ironic.conf and were missing the watcher.yaml which lead to warnings.
As the ironic-conductors don't need the watcher middleware we remove it
from the conductors.