Support legacy config file path and new config file path
With the introduction of IPv6 we will also need a new v6 only config.
That has to be encoded in the file name but also in the name of the
configMap.

Hence, we move the part that checks if the config is present out into a
helper. In that helper, we try the new path, if it is present we return
that path. If not we fall back to the legacy path.

We also give the configMap a naming that is independent of the config
used but basically equals the new config file name without the `.conf`
suffix.