ironic: fix k8s mount configmap issue
The watcher change was not working. K8s was not able to mount the
watcher.conf as the /etc/ironic/ironic.conf.d path was a volume used to
add the secrets for ironic. This worked while manual editing the
deployment. Very big thanks to @I500815 who helped author this.