Nova: Set PYTHONWARNINGS by default
We want to ignore certain warnings from Python, especially the
`SyntaxWarning`, because it gets logged for `pymemcache` seemingly
outside of a `GreenThread` which can lead to services hanging in locks.

We also switch the VMware Hypervisor Deployment to use the same value
`python_warnings` as we already use for any other Deployment.