[Utils] Append urlencode in resolve_secret_urlencode
Using an old vault reference '*vault(path:..)' as input for
resolve_secret_urlencode does not append an urlencode as
the function name might indicate.

During the migration from old vault-injector to the new secret-injector,
we observed this problem with some Neutron passwords. Removing the
urlencode from our Chart (replacing it with a 'resolve_secret_urlencode') lead to an unencoded password.
Simply appending an 'urlencode' back in our Chart does not work either if we migrate
to the new vault reference would
urlencode the string '{{ resolve vault+kvv2:///some/path }}'. Assuming
the secret-injector does not pick this up.