nova/flavors: Add KVM v1 (Cascade-Lake) flavors
As agreed in ADR core/nova/0008 "KVM flavor naming scheme" the KVM
flavors (and eventually, but not in this change, the VMware flavors
too) gain a new hypervisor field after the g/c/m purpose field.
It's "k" for KVM/Qemu (and will be "v" for VMware).

Also the flavors will be suffixed with a "v1" hardware generation
number for Cascade Lake right away.

So, the KVM counterpart for "g_c4_m16" is named "g_k_c4_m16_v1".

A per-region config flag `kvm_enabled` is introduced, defaulting to
false, to allow per-region enablement of KVM flavors.