[rabbitmq] Option to enable SSL for amqp server
The option `enableSsl` configures a certificate with the same name as
the rabbitmq with the prefix `tls-`, which will in turn generate a
secret of the same name.

The only other missing value is the name of the cluster-issuer to
be specified with `certificate.issuerRef.name`

Erlang should automatically reload the certificate, when the file
changes. So, certificate rotation should also be handled.