[unbound] accept both externalIP (singular) and externalIPs (plural)
Keeping externalIP for backwards compatibility's sake.
We might think about retiring it completely later on.

The externalIP value is expected to be a string, i.e. a single IP.
The externalIPs value is expected to be a list.

If both are defined then the values will be merged.
If none are defined then we error out.

Q: Why we need this?
A: There might be a situation where we need to expose unbound over
multiple IPs.