[ironic] Run certrobot twice a day
The prior settings were assuming that we rarely
change the certificates, so no need to run it often.
But we also build up new hardware, and configure
it right away with redfish, so this job becomes
a dependency for the build-up.
Doing it twice a day reduces the time people
have to wait for it to work.