ironic: change traffic policy for pxe-service
Calico qa-de-1 now uses a additional BGPFilter that will surpress service
subnets announcements. Only /32 host routes for Services will be
announced by nodes. Thus breaking TFTP/iPXE traffic. The traffic
would reach the cluster but not the service. We need to add the
`externalTrafficPolicy: local` to make this work.

Increasing the chart version to track this.