[ironic] Move image verification to agent
The ironic-python-agent gets some config values over the
lookup api as 'config' in the json.
And two values are unintuitively taken from the conductor config
section.
So, we have to set them in the api node as well, in order to have the
settings take hold in the agent.

With a current ironic-python-agent, we do not need to validate then
in the conductor itself.