[vcenter-operator] Support secrets-injector
We need to support `secrets-injector` references coming from the values
files. Since some values might be changed already while others may not,
we bump the `utils` chart to the latest version and use its
`resolve_secret` helper on every value that could be a secret. This
takes care of including the proper syntax for the `secrets-injector` if
the secret is a reference.

Since we haven't bumped the `utils` chart in quite some time, we will
also get some other changes. Other than `resolve_secret`, we only use
`helm-toolkit.snippets.kubernetes_metadata_labels` from the `utils`
chart.