[Placement] Support dalmatian release
The only change we need to do is in the policy, which has switched the
rules they use for validation. Since we only ever changed the generic
rules that define an admin, we remove the defaults - they don't match
between yoga and dalmatian and can be easily generated with `tox -e
genpolicy` in Placements repo.

We set both the old (yoga) and the new (dalmatian) generic rules that
define an admin to support smooth upgrading.

We also bump the `appVersion` of the chart to "dalmatian", because
that's the newest we support.