redis: cleanup useless label/templating/validation clutter
Removing the label clutter from the metrics deployment podspec
cannot be done as an in-place edit, so this change renames the
deployment from "redis-metrics" to "redis-exporter" to allow the change
to go through.

This also removes all "required" checks where values.yaml sets a default
value, so the failure branch of "required" cannot actually be taken.