[ironic] move dhparams to secrets
The dhparams were missed when we moved the secrets from configmaps to
secrets. This should make them work with the new secrets-injector. I
also renamed the nginx config from default.conf to the default name
nginx.conf. The emptydir is needed to be able to mount the secret and
possible other configs as well, even when we currently have no other
configs.