[sap-seeds] Filter out flavors without ID
Print them as a comment, so one can cross-check it.

The flavors are often split up over several files by some basic
properties defined by the hardware (i.e. NUMA architecture).
Additionally, we have files that contain additional attributes
for *all* flavors that may be more "policy based" (HANA exclusive
or not, quota sepearate...).

To avoid combinatoric explosion of files, we just want to merge these
to get the combinations by that.
For that, we need to only apply flavors which have their basic
attributes set, and ignore the "policy based" ones.