[ironic] switch curl startup probe to netrc
The curl startup probe used the password in the parameter. That is
broken with the secrets-injector and previously showed the password to
people with kubernetes access. Instead we now use a netrc, in this case
mounted in /etc/ironic instead of the default homedir, and a kubernetes
secret to provide the credentials.