fix(proxysql): ignore metis user
The metis user credentials are not specified in all regions. This results in failures executing this helper.
The user will also never be user with proxysql.
This ignores the user to avoid having to specify the user and grants in the secret values.
Or the need to create user credentials in regions where the service is not running.