[utils] Log system_scope in user logging
When logging the `Context` of a request, we now include the
`system_scope`, too. `oslo.log` defaulted to doing so in
I7982cdb280599fb38362eb837cb527592267b91f, which is part of the `zed`
release. The required changes in `oslo.context` are already part of the
`queens` release.

Citing from `oslo.log`'s commit message:

    As part of Secure RBAC work, support for system users(eg. system admin)
    is being added. However current default format does not include
    the system scope information so we can't determine whether the context
    is non-scoped or system-scoped.