[Neutron] Run rabbitmq credential-helper as sidecar
The RabbitMQ chart offers a credential updater helper, which is executed either as init
or sidecar container. The first option is the default setting (based on
the global values).

Acutally the rabbitmq container never starts as the credential-update
initcontianer never terminates. Seems like it is constantly checking for
credential updates.