[Cinder] Support password ratelimit backend
The Redis helm chart enforces the access via a password. If no
password is set, a random pw is created by container start
automatically.

So far the openstack-ratelimit middleware does not support connecting to
a Redis backend with a password set. Requires https://github.com/sapcc/openstack-rate-limit-middleware/pull/29
The middleware loads the backend password from a file. With this PR we
mount the Redis password into the neutron-server so it will be picked up
by the ratelimit middleware.

reference:
https://github.com/sapcc/helm-charts/pull/8408