[manila] snapmirror capacity for multiple targets
Some setups have more than one snapmirror destination endpoint,
so we multiply now by the number of snapmirror relationships.
But only those that are not under control of manila,
i.e. only if the share server does not follow the naming pattern of
`ma_<UUID>`

And adding an info alert that detects inconsistencies.