[unbound] return NXDOMAIN for blocklisted domains
Now we can blocklist the RFC1918 zones, overriding the hardcoded
local-zone: definitions in the helm chart. We want them to respond
with NXDOMAIN, which is the default behaviour for not enabled AS112
zones. [1]

Going forward, we will remove the zones from the helm chart and
enable them on a case-by-case basis.

We might switch back to REFUSED once the RFC1918 transition takes place.

[1] https://unbound.docs.nlnetlabs.nl/en/latest/manpages/unbound.conf.html#unbound-conf-local-zone