openstack/backup-replication: add .Values.image_refs
When deploying from an OCM bundle, we want for the image refs in the OCM
bundle to take precedence over our usual mirror selection logic based on
`.Values.global.registry` and friends.

I am not sure if this is a pattern that we will recommend to everyone.
For now, backup-replication is our field test for OCM bundles, and we
will see how this pattern feels.