openstack/content-repo: add .Values.image_refs
When deploying from an OCM bundle, we want for the image refs in the OCM
bundle to take precedence over our usual mirror selection logic based on
`.Values.global.registry` and friends.

I am not sure if this is a pattern that we will recommend to everyone.
For now, content-repo and backup-replication are our field test for OCM
bundles, and we will see how this pattern feels.

cf. https://github.com/sapcc/helm-charts/pull/8575
(the same change for openstack/backup-replication)

For alpine-curl, going with `image_refs` switches the imagePullPolicy to
IfNotPresent, because we are not going to use the "latest" tag there.
"latest" is fundamentally incompatible with how `ocm transfer ctf
--copy-resources` copies images: It will refuse to copy the current
"latest" image if a different image with tag "latest" already exists in
the target repo.