postgresql-ng: remove support for .Values.sqlOn{Create,Startup}
The semantics of these fields are called into question with the move
to multitenancy. I tried looking at how these fields are used downstream
to decide how to handle them, but couldn't find anyone using this
feature at all.

This is the first instance of backwards-incompatibility in this feature
branch, but more will come.