postgresql-ng: bump image
This is the first proper image for the v2 series. A weird consequence
of us holding the Dockerfile for this image inside the chart is that we
needed to merge the PR with the content updates first, and then we could
build the image. In hindside, #8644 should not have bumped the chart
version yet...