pgmetrics: change pod name format
This is mostly aesthetics, but since all other postgres pods are called
`$RELEASE-p...`, I think it makes sense to have this be a consistent
pattern to group them together in the sorted output of `k get pods`.

Also, because release names and DB names are usually identical,
we get deployment names like `castellum-castellum-pgmetrics`, which
will be perceived as either silly or a templating bug.
`castellum-pgmetrics-castellum` looks more intentional, and also
follows precedent in secret names like `castellum-pguser-castellum`.