[Neutron] New default logging conf
Introduce new basic logging config for new sapccsentry logger.
Regions requiring different setup can override values in the
secret.vaules.

For most of the prod regions logging config is as most the same as the
default. In some regions the networking_arista logger is configured.
Furthermore in some regions the networking_aci logger has DEBUG log
level and not INFO.