[netapp-exporter] fix broken castellum rules
There are cases where (app, svm, volume) is not unique label set, we
need to add filer to the set.

Also, for the provision_case rules, the group direction is crucial. The
group_left() means the labels are merged from right to left. The right
side must have unique label set. So we put `netapp_volume_labels:manila`
on the right side.