[Neutron] Use old policy setting
With the Neutron Caracal release two things changes with respect to
policies:
1. Per default, Neutron enforces the token scope (e.g. project, domain or system) check.
2. Per default, the old deprecated rules are not evaluated anymore.

With enforcing the new default rules (2) causes PolicyNotAuthorized errors
on our side. To minimize the migration risk, we decided to go on with
the old policy setting (as it currently is) and migrate to the new
defaults in a later step.