[Nova] Disable considering free disk during scheduling
We can easily add more ephemeral disk space, so it's not important to us
during scheduling - it only has to fit, which Placement already takes
care of. This means RAM/VCPUs weight cannot be cancelled out be a lot
more free disk space.

With vastly differing disk space usage of up to 50% this had become an
issue.