Allow users to create projects only in their domains
This change partially restores the policy of project creation. It allows
users to create projects only in their domains. However, it still allows
project creation for certain other admins even across domains.

This is needed because we don't want external users to create projects
everywhere, and we don't want our projects to be created in their
domains.