[Nova] api-metadata multi-secret with auto-restart
nova-api-metadata pods will automatically restart through reloader, when
the newly-introduce nova-api-metadata k8s Secret changes. This Secret
contains the shared metadata secret Neutron also knows.

It should be safe to restart nova-api-metadata on Vault secret change,
because it's a multi-replica service. Changing MariaDB/RabbitMQ secrets
requires an explicit rollout, so these wouldn't be picked up with the
restart. The only change it could pick, is Keystone credentials, but
those won't change for a while and even then switch to application
credentials and by that should support old/new dual-credentials like
with MariaDB/RabbitMQ and what we try to build here.

We also support setting multiple secrets here, because Nova gets support
for that in our fork. This enables smoother secret rotation. Since the
current settings do not use a list but only a string, we had to add some
code used during the migration, so we can safely roll this change out
without having the values change in place.