[openstack/nova] Nanny for orphaned migrations
Since we should only have few migrations ongoing, the query should be
quite fast. We trigger it often as it will be a blocker to re-installing
kvm hosts, as they get decomissioned each time, but have to wait for no
migrations being associated with the compute-host.
Disabled for now, so we can activate it selectively in some regions for
now.