limes: reduce cost of some pgmetrics queries
Using the new `cluster_resources.path` field to save an extra join.

Both those queries are reported in the eu-de-1 postgresql.log as taking
very long (on the order of 3-5 seconds for the overspent-quota
query and 300-500 ms for the free-swift-quota query).

Despite the removed table (`cluster_services`) being very small, this
somehow appears to save a shocking amount of cost. I can only test this
in QA right now, but the overspent-quota query went down from 778 -> 637
ms and the free-swift-quota query went from 118 -> 10 ms there.