[ironic] increase parallelism for cleaning
With max_concurrent_clean it's possible to set the parallelism level for
cleaning. By default it's 50 but we have HDA who work on 50 nodes in
parallel, so a single node that was in cleaning before might break their
deployment. So we give a buffer of 10, which should hopefully be enough.
As we have multiple conductors it might be safe to increase it further.
With the current setup no conductor will have more nodes than around 28.