neutron-netns-prober: include network_project_id in alert
* The neutron-netns-prober is the refactored version of the ns-exporter.
  The former is currently only deployed on qa and does not hit
production regions. In this commit we include the label
network_project_id so that it shows in the alert. Notice: there is only
only project_id per network. This doesn't hold for the first hour where
the exporter is deployed. In this first hour there will be two matching
metrics (one where network_project_id is empty and another where is not.
Since the versions have been deployed some hours ago this shouldn't be a
problem).