ns-exporter: include network_project_id in alert
* The ns-exporter is the monitoring tool that is currently deployed on
  production regions (and differs from the neutron-netns-prober in that
it still has the old code base). The ns-prober is deployed in production
regions.
* In this commit we include the label network_project_id so that it
  shows in the alert. Notice: there is only only project_id per network.
This doesn't hold for the first hour where the exporter is deployed. In
this first hour there will be two matching metrics (one where
network_project_id is empty and another where is not.  Since the
versions have been deployed some hours ago this shouldn't be a problem).