[cinder] Split api in external and internal pods
We want to configure different rate-limits and policies for
inter-service communication compared to user communication. To make this
work, we need different cinder-api pods. Keystone's service-catalog
already supports having different endpoints for these purposes.

We spawn a cinder-api-external and a cinder-api-internal deployment each
with their own cinder-api-external/internal-ratelimit-redis, since the
rate-limit-middleware doesn't support re-using the same redis for
multiple apis and it allows us to deploy the redis next to the
cinder-api in the same cluster, if we want to.

Since most of the configuration/YAML is the same between the two
cinder-api pods, we define template function and provide the individual
configuration as parameters.

During the rollout, there will be a time where the new internal endpoint
is not yet synced into the Keystone catalogue. Therefore, we also keep
the old "cinder-api" Service alive, pointing to the cinder-api-external
Deployment as before.

It should be easy to extend this pattern with an "admin" API endpoint as
supported by Keystone's catalogue if we choose to use that in the
future.

This commit contains some restructuring, e.g. introducing new
cinder-api-external-etc and cinder-api-internal-etc ConfigMaps to hold
and template API configuration and switching to projected volume
mounting in the api Deployment for better readability and static config
during the runtime.