[ironic] remove kos node update
We no longer need the node update. All nodes are already switched. But
we might now want to switch settings for some nodes for example for
secure boot use cases or in the future to switch all nodes to the
generic redfish driver and we don't need the kos to change nodes back
the whole time.