[neutron] Configure Metadata Service Query Rate-limiting
This change introduces a value to our neutron deployment allowing to
turn on and configure the rate-limiting for the metadata service running
as part of our neutron network agents. It's turned off by default but
can be toggled on by providing the respective values during a helm
release.

For more information about it's capabilities and configuration, refer to
https://docs.openstack.org/neutron/latest/admin/config-metadata-rate-limiting.html