[Cinder] Disable rate-limiting
We disable rate-limiting, because this leads to inconsistencies with
volume-attachments. Even if the first requests Nova does on behalf of
the user work, later requests can fail because of rate-limiting and
leave the volume in an intermediate state until fixed manually.

We're working on a better solution already, but since we do not have to
rely on rate-limiting currently, want to reduce our manual efforts and
increase customer happiness by disabling rate-limiting again.