[nova] Track recently retried migrations
Ideally, a failed migration should not let the VM end up in error state,
so an operator might be inclined to retry the operation. This gauge
keeps track if that happens more often, and how often, so one might
identify issues with specific VMs or hosts.

The interval chosen of 15 minutes is fairly arbitrary. The main point
is to limit the history to keep the query hopefully reasonably quick.