[Nova] Fix shellinabox lua mount path
The path /app is not in the search paths of lua when running with
openresty. To fix the shellinabox console, we have to mount the lua
files into /usr/local/openresty/site/lualib to make them importable.
It's the same path we also use in the nova-console-mks Deployments,
which are also based on openresty.