Add option to be able to enable cinder's image cache
This adds the ability to enable cinder's build in image cache
mechanism.

TODO: need to figure out the secrets injection for a cinder
tenant project_id and cinder tenant user_id.

Need to figure out the mess that is _volume-secrets.conf.tpl
to add the ability to enable the image cache for a specific
backend only.