openstack/kvm-cinder-netapp-operator adapt fields for changed CRDs (#10194)
To support multiple NetappFilerWatcher resources coexisting, the volumeregextemplate field was moved there from CinderNetappConfig.

To support assisted migration of volumes across filers in Cinder, NetappFiler resources need to know which secrets of filers to consider for Cinder configuration. This and a few other previously missing fields got added to the template.