[openstack/autoclean] Cronjobs to cleanup our staging environment
It will mark every week all instances with tags, and if the tags are
not removed, it is assumed that the resource can be deleted after two
weeks. After one week, the instances are getting stopped first to make
to give people some time to react to it.

Obviously, that should only run in staging, as it works on the whole
region.