[neutron] Allow users to create subnets w/o DHCP
Due to the old policy users cannot create a subnet with DHCP disabled,
but they can disable this in an update call. As the policy is pretty old
and users can do it anyway (and there's no problem with users
configuring their networks this way) we're now allowing DHCP to be
disabled on creation.