[nova] increase default vm size threshold on smal hvs to 32g
A lot of our smallerr HVs don't get utilitzed as much as they could
since they are limited to 16G VMs. A select few regions are an exception
and enable 32G or 64G VMs already. Change the default aross all regions
to 32G now with the option to go higher later if required.